# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.BindFileRequest:
    properties:
      entity_id:
        type: string
      entity_type:
        type: string
    required:
    - entity_id
    - entity_type
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.FileResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      entity_id:
        type: string
      entity_type:
        type: string
      file_name:
        type: string
      file_type:
        type: string
      id:
        type: string
      mime_type:
        type: string
      original_name:
        type: string
      size:
        type: integer
      size_readable:
        type: string
      storage_type:
        type: string
      uploader_id:
        type: string
      url:
        type: string
    type: object
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.FileStatsResponse:
    properties:
      audio_count:
        type: integer
      audio_size:
        type: integer
      doc_count:
        type: integer
      doc_size:
        type: integer
      image_count:
        type: integer
      image_size:
        type: integer
      total_count:
        type: integer
      total_size:
        type: integer
      video_count:
        type: integer
      video_size:
        type: integer
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_internal_application_dto.UploadResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.FileResponse'
        type: array
    type: object
...
  github_com_Snowitty-Re_CNtunyuan_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cntunyuan.org
    name: CNtunyuan Team
    url: https://github.com/Snowitty-Re/CNtunyuan
  description: 团圆寻亲志愿者系统后端 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Snowitty-Re/CNtunyuan
  title: 团圆寻亲志愿者系统 API
  version: 1.0.0
paths:
...
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传单个文件，支持图片、音频、视频等格式，文件大小限制50MB
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.FileResponse'
              type: object
        "400":
          description: 请求参数错误（未提供文件或文件过大）
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 单文件上传
      tags:
      - 文件管理
  /upload/{id}:
    delete:
      consumes:
      - application/json
      description: 根据文件ID删除文件（物理删除文件及数据库记录）
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 删除成功，无返回内容
        "400":
          description: 请求参数错误（文件ID不能为空）
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除文件
      tags:
      - 文件管理
    get:
      consumes:
      - application/json
      description: 根据文件ID获取文件详细信息（元数据，不包含文件内容）
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.FileResponse'
              type: object
        "400":
          description: 请求参数错误（文件ID不能为空）
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取文件详情
      tags:
      - 文件管理
  /upload/{id}/bind:
    put:
      consumes:
      - application/json
      description: 将已上传的文件绑定到指定实体（如走失人员、方言等）
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      - description: 绑定请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.BindFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 绑定文件到实体
      tags:
      - 文件管理
  /upload/{id}/download:
    get:
      consumes:
      - application/json
      description: 根据文件ID下载文件内容，返回二进制文件流
      parameters:
      - description: 文件ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件内容
          schema:
            type: file
        "400":
          description: 请求参数错误（文件ID不能为空）
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 下载文件
      tags:
      - 文件管理
  /upload/batch:
    post:
      consumes:
      - multipart/form-data
      description: 同时上传多个文件，支持图片、音频、视频等格式
      parameters:
      - description: 要上传的文件列表（支持多个）
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.UploadResponse'
              type: object
        "400":
          description: 请求参数错误（未提供文件）
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 批量文件上传
      tags:
      - 文件管理
  /upload/entity/{type}/{id}:
    get:
      consumes:
      - application/json
      description: 根据实体类型和ID获取关联的所有文件列表
      parameters:
      - description: 实体类型（如：missing_person, dialect, user等）
        in: path
        name: type
        required: true
        type: string
      - description: 实体ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.FileResponse'
                  type: array
              type: object
        "400":
          description: 请求参数错误（实体类型或ID不能为空）
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取实体关联的文件列表
      tags:
      - 文件管理
  /upload/stats:
    get:
      consumes:
      - application/json
      description: 获取系统文件统计数据，包括总数量、总大小、各类型文件统计等
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_application_dto.FileStatsResponse'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取文件统计信息
      tags:
      - 文件管理
...
securityDefinitions:
  Bearer:
    description: 请输入 JWT Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"